Title: 최솟값 3개, Time: 1045ms, Memory: 23MB, Status: Passed - CodeTree

+++ 241006/최솟값 3개/top-3-smallest-number.java
@@ -28,7 +28,7 @@ public class Main {
             pq.add(num2);
             pq.add(num3);
 
-            System.out.println(num1 * num2 * num3);
+            System.out.println((long) num1 * num2 * num3);
         }
     }
